moodpulse-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
transformers==4.46.3
torch==2.4.1
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
huggingface-hub==0.26.2
numpy==1.26.4
accelerate==1.0.1